+++ src/Components/Join/signInUp.css
@@ -1,0 +1,162 @@
+.card {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  min-height: 55vh;
+  width: 600px;
+  background-color: #080312;
+  color: white;
+  border-radius: 4px;
+  gap: 1rem;
+  transform-style: preserve-3d;
+  transition: 1000ms;
+  transform: perspective(1000px) rotateY(var(--rotate-y, 0))
+  translateY(var(--translate-y,0));
+ padding: 1rem;
+}
+.card.flip{
+  display: flex;
+  --rotate-y: 180deg;
+}
+.card .front{
+  left:0;
+}
+.front input, .back input{
+  width: 90%;
+  height: 48px;
+  background-color: #1f2029;
+  border-radius: 4px;
+  text-align: center;
+  font-size: 14px;
+  line-height: 22px;
+  letter-spacing: 0.5px;
+  outline: none;
+  color: #c4c3ca;
+  background-color: #1f2029;
+  border: none;
+  -webkit-transition: all 200ms linear;
+  transition: all 200ms linear;
+  box-shadow: 0 4px 8px 0 rgba(21, 21, 21, 0.2);
+  margin: 0.5rem;
+  
+}
+.card .front, .card .back{
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 1rem;
+  position: absolute;
+  padding: 1rem;
+  backface-visibility: hidden;
+
+}
+.card .front h4, .card .back h4{
+  text-transform: uppercase;
+}
+
+.card .back input{
+  width: 45%;
+  
+}
+
+.card .back{
+  transform: rotateY(180deg);
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+  align-items: center;
+}
+
+.full-height {
+
+  width: 100%;
+  min-height: 65vh;
+  background-image: url("https://i.gifer.com/origin/e2/e2aec645e3f805bfeef5468bc9bf3a34.gif");
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+.content{
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background: rgba(255, 255, 255, 0.2);
+  backdrop-filter: blur(5px) sepia(.5); 
+  min-height: 65vh;
+	width: 100%;
+  flex-grow: 1;
+  gap: 1rem;
+  padding: 2rem 0;
+}
+ .btn {
+  border-radius: 4px;
+  height: 44px;
+  font-size: 13px;
+  font-weight: 600;
+  text-transform: uppercase;
+  -webkit-transition: all 200ms linear;
+  transition: all 200ms linear;
+  padding: 0 30px;
+  letter-spacing: 1px;
+  display: -webkit-inline-flex;
+  display: -ms-inline-flexbox;
+  display: inline-flex;
+  -webkit-align-items: center;
+  -moz-align-items: center;
+  -ms-align-items: center;
+  align-items: center;
+  -webkit-justify-content: center;
+  -moz-justify-content: center;
+  -ms-justify-content: center;
+  justify-content: center;
+  -ms-flex-pack: center;
+  text-align: center;
+  border: none;
+  background-color: #B75BCF;
+  color: #102770;
+  box-shadow: 0 8px 24px 0 rgba(255, 235, 167, 0.2);
+} 
+.btn:active,
+.btn:focus {
+  background-color: #102770;
+  color: #ffa7f8;
+  box-shadow: 0 8px 24px 0 rgba(16, 39, 112, 0.2);
+} 
+.btn:hover {
+  background-color: #102770;
+  color: #ffa7f8;
+  box-shadow: 0 8px 24px 0 rgba(16, 39, 112, 0.2);
+} 
+
+
+/* <-------- media quary --------> */
+@media (max-width: 768px){
+  .card{
+    width: 90%;
+    min-height: 400px;
+  }
+}
+
+@media (max-width:425px){
+  .card{
+    height: 100vh;
+  }
+  .card .back input{
+    width: 90%;   
+  }
+
+ 
+}
+/* .front{
+  
+} */
+ @media (min-width:2000px) {
+  .btn, .card, .front input, .back input{
+  font-size: 2rem;
+ } 
+}
