+++ src/Components/Navigation/NavBar.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import NavLinks from './NavLinks'
+import './navBar.css'
+
+export default function NavBar() {
+  return (
+    <nav className="content-buttom-ancor">
+    <NavLinks/>
+    </nav>
+  )
+}
