+++ src/Components/Navigation/mobileNav.css
@@ -1,0 +1,46 @@
+.mobileNav {
+    display: flex;
+    flex-direction: row;
+    min-height: 5vh;
+    width: 100%;
+
+} 
+
+.menu{
+    padding-top: 1rem;
+    gap:.5rem;
+    display: flex;
+    flex-direction: column; 
+    width: 100%;
+    min-height: 10vh;
+    background: #181818;
+    text-decoration: none;
+    color: #00fffb;
+    margin-right: .5rem;
+  
+  }
+  .menu a{
+    text-decoration: none;
+  }
+  .menu a:hover  {
+    color: white;
+  }
+  
+.mobileNav{
+    display: none;
+}
+.mobileNav img{
+    position: absolute;
+    top: 50px;
+    right: 0;
+    margin-right: 1rem;
+  width: 30px;
+  height: 30px;
+}
+
+@media (max-width: 768px){
+    .mobileNav{
+     display: flex;
+            
+    }
+}
