+++ src/Components/PhotoHome.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "../Pages/home.css";
 import { Fade } from "react-awesome-reveal";
-import { height } from "dom7";
+// import { height } from "dom7";
 
 export default function PhotoHome(props) {
   let { name, photo } = props;
