+++ src/Components/Switch/switch.css
@@ -1,0 +1,61 @@
+.switch-container{
+    display: flex;
+    min-width: 20%;
+    height: 50px;
+    gap:.5rem;
+    align-items: center;
+    justify-content: center;
+    background-color: rgba(189, 43, 226, 0.436);
+    padding: .8rem;
+    border-radius: 100px;
+    transition: background-color .2s;
+}
+.tittle{
+    color: white;
+
+}
+.react-switch-checkbox {
+    height: 0;
+    width: 0;
+    visibility: hidden;
+  }
+   
+  .react-switch-label {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    cursor: pointer;
+    width: 90px;
+    height: 40px;
+    background: rgba(200, 51, 200, 0.449);
+    border-radius: 100px;
+    position: relative;
+    transition: background-color .2s;
+  }
+   
+  .react-switch-label .react-switch-button {
+
+    content: '';
+    position: absolute;
+    top: .8px;
+    left: 2px;
+    width: 45px;
+    height: 40px;
+    border-radius: 45px;
+    transition: 0.2s;
+    background: #fff;
+    box-shadow: 0 0 2px 0 rgba(10, 10, 10, 0.29);
+  }
+   
+  .react-switch-checkbox:checked + .react-switch-label .react-switch-button {
+    left: calc(100% - 2px);
+    transform: translateX(-100%);
+  }
+   
+  .react-switch-label:active .react-switch-button {
+    width: 60px;
+  }
+  .active{
+    color: rgba(85, 8, 85, 0.938);
+    font-weight: 800;
+  }
