+++ src/layout/Layout.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Header from './header/Header'
-import { Footer } from './footer/Footer'
+import Header from './Header/Header'
+import { Footer } from './Footer/Footer'
 
 export default function Layout(props) {
   return (
